label stations

+++ app/src/main/java/com/example/ecoroute/ui/home/HomeFragment.kt
@@ -11,7 +11,6 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.ProgressBar
-import androidx.annotation.Keep
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
@@ -23,8 +22,8 @@ import com.example.ecoroute.utils.ApplicationUtils
 import com.example.ecoroute.utils.URLBuilder
 import com.example.ecoroute.utils.UiUtils
 import com.google.android.material.snackbar.Snackbar
+import com.google.android.material.switchmaterial.SwitchMaterial
 import com.mapbox.geojson.Point
-import com.mapbox.maps.Image
 import com.mapbox.maps.MapView
 import com.mapbox.maps.MapboxMap
 import com.mapbox.maps.Style
@@ -44,7 +43,6 @@ import com.mapbox.navigation.ui.maps.camera.data.MapboxNavigationViewportDataSou
 import com.mapbox.navigation.ui.maps.camera.lifecycle.NavigationBasicGesturesHandler
 import com.mapbox.navigation.ui.maps.location.NavigationLocationProvider
 import kotlinx.android.synthetic.main.fragment_home.*
-import kotlinx.android.synthetic.main.fragment_user.*
 
 
 @SuppressLint("LogNotTimber", "StringFormatInvalid", "SetTextI18n")
@@ -58,7 +56,8 @@ class HomeFragment : Fragment() {
     private val viewModel: HomeViewModel by viewModels()
     private lateinit var pb: ProgressBar
     private val ctx = ApplicationUtils.getContext()
-
+    private var station_response : ArrayList<NearbyStationsResponse.NearbyStationsResponseItem>? = null
+    private lateinit var ms : SwitchMaterial
 
     private lateinit var homeMapView: MapView
     private lateinit var homemapboxMap: MapboxMap
@@ -84,6 +83,7 @@ class HomeFragment : Fragment() {
         pb = root.findViewById(R.id.pbHome)
         homeMapView = root.findViewById(R.id.mapViewHome)
         homemapboxMap = homeMapView.getMapboxMap()
+//        ms = root.findViewById(R.id.material_switch)
 
         annotationApi = homeMapView.annotations
         pointAnnotationManager = annotationApi.createPointAnnotationManager()
@@ -95,7 +95,13 @@ class HomeFragment : Fragment() {
                 Snackbar.LENGTH_SHORT
             ).show()
         }
-
+//        ms.setOnCheckedChangeListener { _, isChecked ->
+//            if (isChecked) {
+//                markStations(station_response, true)
+//            } else {
+//                markStations(station_response, false)
+//            }
+//        }
         onMapReady()
 
 
@@ -161,12 +167,13 @@ class HomeFragment : Fragment() {
 
             if (viewModel.successful.value != null) {
 
+                station_response = mResponse
                 pb.visibility = View.INVISIBLE
                 FINDING_STATION = false
                 uiUtilInstance.showToast(ctx, viewModel.message.value.toString())
                 Log.e(TAG, mResponse.toString())
                 if (viewModel.successful.value == true) {
-                    markStations(mResponse)
+                    markStations(mResponse, true)
                 }
 
             } else {
@@ -175,17 +182,21 @@ class HomeFragment : Fragment() {
         })
     }
 
-    private fun markStations(mResponse: ArrayList<NearbyStationsResponse.NearbyStationsResponseItem>) {
+    private fun markStations(mResponse: ArrayList<NearbyStationsResponse.NearbyStationsResponseItem>?, show : Boolean) {
+
+        if(mResponse == null){
+            return
+        }
 
         Log.e(TAG, "Marking ${mResponse.size} stations")
         pointAnnotationManager.deleteAll()
         for (stations in mResponse) {
-            attachMarkers(stations.position.lat, stations.position.lon)
+            attachMarkers(stations.position.lat, stations.position.lon, stations.poi.name, show)
         }
 
     }
 
-    private fun attachMarkers(_latitude: Double, _longitude: Double) {
+    private fun attachMarkers(_latitude: Double, _longitude: Double, name: String, show: Boolean) {
 
 
         uiUtilInstance.bitmapFromDrawableRes(
@@ -197,7 +208,12 @@ class HomeFragment : Fragment() {
                     _longitude, _latitude
                 )
             ).withIconImage(it).withTextAnchor(TextAnchor.TOP)
+            if(show){
 
+                pointAnnotationOptions.textField = name
+                pointAnnotationOptions.textSize= 8.0
+
+            }
             pointAnnotationManager.create(pointAnnotationOptions)
